requirements.txt
loguru==0.6.0
python-telegram-bot==21.10
python_okx==0.3.5
Requests==2.32.3
